project 3/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
requests==2.31.0
openai==1.13.3
groq==0.4.2
httpx==0.27.0